backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
motor==3.4.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.9
python-socketio==5.11.2
cloudinary==1.40.0
qrcode==7.4.2
Pillow==10.4.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
